Changed URL routings and added a product page
URL routings were interfering with eachother. As
categories and products both started with a slug,
so when trying to access a product page, the urls
tried to route the product as a category.

E.g if bikezon/bike is a product, the url was routing
it as a category, which did not exist, hence showed an error.

Catgeories now start with category/ (e.g category/parts/handles)
and products start with product/ (e.g product/dirt-bike-200)

Both now work.