Added an option to print email to stdout
THe email to reset password requires a gmail account with
creds. Incase someone does not want to use a gmail account
or we are not willing to supply our creds, if the email.json
file is not found in BASE_DIR/email.json the application
will instead print the email to stdout (terminal) instead
of sending it.

To configure the application to use an email, supply it gmail
creds in the email.json file, e.g:

email.json
{
    "email": "your.email@gmail.com",
    "password": "your_gmail_password"
}

**Make sure to not commit/push your gmail creds!!!**