Fixed the URLS
**PLEASE READ THIS BEFORE CHANGING URLS AGAIN**

Fixed URLS because slugs in start of both were causing conflicts.
If we have slug/... for category and slug/.. for products then
the urls get mixed up deciding if we are trying to access a
category or a product.
E.G:
If there exists a product bike1 but no category bike1 then the app
would try to find the CATEGORY slug bike1 and error because it
doesn't exist. There are other issues with that implementation too.

ADDITIONALLY it makes sense to have catgeory/... and product/..
for organisation.

Tested, everything works now.

Expanded the product page a bit to include a template of contacting the
seller and a product pic. Everything needs major front-end work.